checkpoint conversion to new trait
